Updated main.cpp (debugged and working ID scan)
This version has been debugged and works when scanning ID's. When button is pressed the rfid module will detect a card if it is in range. once detected it will be scanned and the serial number (tag) will be printed to the serial monitor. 

Next steps:
Implement/ debug bluetooth.
Implement some sort of way to have a period after clicking the button where it scans for a card instead of instant scan.
anything else we come up with...